Version 1.0.2
En esta revision ya se corrigio el problema de la actualizacion de la vista en todas la secciones de la app, dejando las vistas actualizadas en todo momento.

Tambien ya se incluye la etapa de la adicion de notas en cada categoria respectiva.

Se implemento la eliminacion de una nota, en caso de que este nota se encuentre en una categoria que no sea papelera, esta se manda directamente a la papelera antes de su eliminacion deifinitiva. Si se encuentra ya en la papelera su eliminacion es completa.
En caso de que se pretenda hacer una adicion en la categoria de papelera, el sistema prohibe dicha accion.

En caso de que se elimine una categoria, todas las notas que se encuentran dentro de la misma se van directamente a la categoria "Sin categoria".

El sistema no permite eliminar las categorias de papelera ni de Sin categoria.

Tambien ya se gestionan los permisos necesarios para poder tener acceso a la galeria y poder escribir la imagen en nuestra ruta de almacenamiento.

Faltantes:

Permitir la edicion de las notas en tiempo real.

Agregar las notas multimedia.

La implementacion con la camara del celular

+++ app/src/main/java/com/it/save/saveit/sqlite/model/CategoriaAdministradora.java
@@ -1,10 +1,9 @@
 package com.it.save.saveit.sqlite.model;
 
 import android.content.Context;
-import android.widget.Toast;
 
+import com.it.save.saveit.AdaptadorCategorias;
 import com.it.save.saveit.CategoriaClass;
-import com.it.save.saveit.MainActivity;
 import com.it.save.saveit.R;
 import com.it.save.saveit.sqlite.helper.DatabaseHelper;
 
@@ -16,12 +15,16 @@ import java.util.Vector;
 
 public class CategoriaAdministradora
 {
+    private Vector<CategoriaClass> categorias;
+    private AdaptadorCategorias adaptador;
     public DatabaseHelper db;
     Context aux;
     public CategoriaAdministradora(Context context)
     {
         db = new DatabaseHelper(context);
         aux=context;
+        categorias=CargaLibros();
+        adaptador=new AdaptadorCategorias(context,categorias);
     }
     public Vector<CategoriaClass> CargaLibros()
     {
@@ -53,13 +56,29 @@ public class CategoriaAdministradora
         db.insertarCategoria(categoria);
         db.closeDB();
     }
-    public void EliminaCategoria(int pos)
+    public String EliminaCategoria(int pos)
     {
-        db.deleteCategoria(CargaLibros().get(pos).getCategoria());
-        db.closeDB();
+        String categ =CargaLibros().get(pos).getCategoria();
+        if(!categ.equals("Sin categoria") && !categ.equals("Papelera de reciclaje"))
+        {
+            db.updateCategoriaNota(categ);
+            db.deleteCategoria(CargaLibros().get(pos).getCategoria());
+            db.closeDB();
+        }
+        else
+            return "La categoria no se puede eliminar";
+        return "Eliminado exitosamente";
     }
     public String getCategoria(int pos)
     {
         return CargaLibros().get(pos).getCategoria();
+    }
+
+    public Vector<CategoriaClass> getCategorias() {
+        return categorias;
+    }
+    public AdaptadorCategorias getAdaptador()
+    {
+        return adaptador;
     }
 }
